websocket direct for drawing data

+++ src/main/java/com/csse433/blackboard/message/dto/InboundDrawingDto.java
@@ -18,9 +18,9 @@ public class InboundDrawingDto {
     @Data
     public static class Content {
         private String color;
-        private int y1;
-        private int x1;
-        private int y0;
-        private int x0;
+        private double y1;
+        private double x1;
+        private double y0;
+        private double x0;
     }
 }
